feat : 오늘의 메뉴 안내에서 평가 추가
- 오늘의 메뉴 안내 시 평가 결과 안내
- 평가 함수는 evaluate.js 로 분리
- evaluate.js 는 메뉴를 받아 해당 메뉴를 평가하여 별점을 리턴
